+++ crypto_sign/cross/clean/keccakf1600.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 void PQCLEAN_CROSS_CLEAN_KeccakF1600_StateExtractBytes(uint64_t *state, unsigned char *data,
-                                   uint16_t offset, uint16_t length);
+                                   uint32_t offset, uint32_t length);
 void PQCLEAN_CROSS_CLEAN_KeccakF1600_StateXORBytes(uint64_t *state, const unsigned char *data,
-                               uint16_t offset, uint16_t length);
+                               uint32_t offset, uint32_t length);
 void PQCLEAN_CROSS_CLEAN_KeccakF1600_StatePermute(uint64_t *state);
